+++ app/src/main/java/com/example/rehabapp/EmgFrame.kt
@@ -3,20 +3,20 @@ package com.example.rehabapp
 class EmgFrame(dataArray: UByteArray) {
 
     private var dataArray = UByteArray(4)
-    private var rawEmg: Float = 0.0F
-    private var filtredEmg: Float = 0.0F
+    private var rawEmg: Int = 0
+    private var filtredEmg: Int = 0
 
     init {
         this.dataArray=dataArray
-        rawEmg = ((dataArray[0].toInt() shl 8) or dataArray[1].toInt()).toFloat()
-        filtredEmg = ((dataArray[2].toInt() shl 8) or dataArray[3].toInt()).toFloat()
+        rawEmg = (dataArray[0].toInt() shl 8) or dataArray[1].toInt()
+        filtredEmg = (dataArray[2].toInt() shl 8) or dataArray[3].toInt()
     }
 
-    fun getRawEmg(): Float{
+    fun getRawEmg(): Int{
         return rawEmg
     }
 
-    fun getFiltredEmg(): Float{
+    fun getFiltredEmg(): Int{
         return filtredEmg
     }
